+++ cli/cmd_slq_test.go
@@ -8,6 +8,8 @@ import (
 	"strings"
 	"testing"
 
+	"github.com/neilotoole/sq/cli"
+
 	"github.com/neilotoole/sq/cli/testrun"
 
 	"github.com/stretchr/testify/assert"
@@ -19,7 +21,7 @@ import (
 	"github.com/neilotoole/sq/testh/sakila"
 )
 
-// TestCmdSLQ_Insert tests "sq QUERY --insert=@src.tbl".
+// TestCmdSLQ_Insert_Create tests "sq QUERY --insert=@src.tbl".
 func TestCmdSLQ_Insert_Create(t *testing.T) {
 	th := testh.New(t)
 	originSrc, destSrc := th.Source(sakila.SL3), th.Source(sakila.SL3)
@@ -191,4 +193,84 @@ func TestCmdSLQ_ActiveSrcHandle(t *testing.T) {
 	require.NoError(t, err)
 	recs = tr.MustReadCSV()
 	require.Equal(t, sakila.TblActorCount, len(recs))
+}
+
+func TestPreprocessFlagArgVars(t *testing.T) {
+	testCases := []struct {
+		name    string
+		in      []string
+		want    []string
+		wantErr bool
+	}{
+		{
+			name: "empty",
+			in:   []string{},
+			want: []string{},
+		},
+		{
+			name: "no flags",
+			in:   []string{".actor"},
+			want: []string{".actor"},
+		},
+		{
+			name: "non-arg flag",
+			in:   []string{"--json", ".actor"},
+			want: []string{"--json", ".actor"},
+		},
+		{
+			name: "non-arg flag with value",
+			in:   []string{"--json", "true", ".actor"},
+			want: []string{"--json", "true", ".actor"},
+		},
+		{
+			name: "single arg flag",
+			in:   []string{"--arg", "name", "TOM", ".actor"},
+			want: []string{"--arg", "name:TOM", ".actor"},
+		},
+		{
+			name:    "invalid arg name",
+			in:      []string{"--arg", "na me", "TOM", ".actor"},
+			wantErr: true,
+		},
+		{
+			name:    "invalid arg name (with colon)",
+			in:      []string{"--arg", "na:me", "TOM", ".actor"},
+			wantErr: true,
+		},
+		{
+			name: "colon in value",
+			in:   []string{"--arg", "name", "T:OM", ".actor"},
+			want: []string{"--arg", "name:T:OM", ".actor"},
+		},
+		{
+			name: "single arg flag with whitespace",
+			in:   []string{"--arg", "name", "TOM DOWD", ".actor"},
+			want: []string{"--arg", "name:TOM DOWD", ".actor"},
+		},
+		{
+			name: "two arg flags",
+			in:   []string{"--arg", "name", "TOM", "--arg", "eyes", "blue", ".actor"},
+			want: []string{"--arg", "name:TOM", "--arg", "eyes:blue", ".actor"},
+		},
+		{
+			name: "two arg flags with interspersed flag",
+			in:   []string{"--arg", "name", "TOM", "--json", "true", "--arg", "eyes", "blue", ".actor"},
+			want: []string{"--arg", "name:TOM", "--json", "true", "--arg", "eyes:blue", ".actor"},
+		},
+	}
+
+	for _, tc := range testCases {
+		tc := tc
+		t.Run(tc.name, func(t *testing.T) {
+			got, gotErr := cli.PreprocessFlagArgVars(tc.in)
+			if tc.wantErr {
+				t.Log(gotErr.Error())
+				require.Error(t, gotErr)
+				return
+			}
+
+			require.NoError(t, gotErr)
+			require.EqualValues(t, tc.want, got)
+		})
+	}
 }